revert: thunk - move
